% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_level.R
\name{extract_levels}
\alias{extract_levels}
\title{Extract specific levels from a multiscale \code{ome-zarr} object}
\usage{
extract_levels(x, levels)
}
\arguments{
\item{x}{An \code{ome-zarr} object.}

\item{levels}{Integer vector specifying the levels to extract.}
}
\value{
\itemize{
\item If \code{levels} is of length 1, an array
\item If \code{levels} is of length more than 1, an \code{ome-zarr} object
}

An object of \code{ome_zarr} (OME-Zarr) class representing an
image or label pyramid.
}
\description{
Extract specific levels from a multiscale \code{ome-zarr} object
}
\examples{
omezarrzip <- system.file("extdata",
                          "test_ngff_image_v04.ome.zarr.zip",
                          package = "rome")
dir.create(td <- tempfile())
unzip(omezarrzip, exdir = td)
x <- ome_read(td)
extract_levels(x, c(1, 3))
extract_levels(x, 2)

}
