% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ome_validate.R
\name{ome_validate}
\alias{ome_validate}
\title{Validate a multiscale OME-Zarr file}
\usage{
ome_validate(path, s3_client = NULL)
}
\arguments{
\item{path}{Path to the OME-Zarr file.}

\item{s3_client}{Object created by \code{\link[paws.storage:s3]{paws.storage::s3()}}. Only required for a
file on S3. Leave as \code{NULL} for a file on local storage.}
}
\value{
This function is used for its side-effect and will return the type of the
OME-Zarr schema (image, label), otherwise will invoke an error when
passed an invalid OME-Zarr file
}
\description{
Validate a multiscale OME-Zarr file
}
\examples{
omezarrzip <- system.file("extdata",
                          "test_ngff_image_v04.ome.zarr.zip",
                          package = "rome")
dir.create(td <- tempfile())
unzip(omezarrzip, exdir = td)
ome_validate(td)
}
