% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ome_write.R
\name{ome_write}
\alias{ome_write}
\alias{ome_write,Image-method}
\alias{ome_write,character-method}
\alias{ome_write,array-method}
\title{ome_write}
\usage{
\S4method{ome_write}{Image}(
  image,
  path = "/",
  axes = NULL,
  scalefactors = c(2, 2, 2, 2),
  version = c("0.4", "0.5"),
  storage_options = NULL,
  type = c("image", "label")
)

\S4method{ome_write}{character}(
  image,
  path = "/",
  axes = NULL,
  scalefactors = c(2, 2, 2, 2),
  version = c("0.4", "0.5"),
  storage_options = NULL,
  type = c("image", "label")
)

\S4method{ome_write}{array}(
  image,
  path = "/",
  axes = NULL,
  scalefactors = c(2, 2, 2, 2),
  version = c("0.4", "0.5"),
  storage_options = NULL,
  type = c("image", "label")
)
}
\arguments{
\item{image}{an n-dimensional (or a path to an) array representing the image data (1<n<6)}

\item{path}{the path to writing ome.zarr}

\item{axes}{a character vector specifying the axes of the image
(e.g. c("t", "c", "z", "y", "x"))}

\item{scalefactors}{Scale factors to apply to construct a multiscale image.
Importantly, each scale factor is relative to the previous scale factor.
For example, if the scale factors are c(2, 2, 2), the returned multiscale
image will have 4 scales.}

\item{version}{OME-ZARR version (0.4 or 0.5), lower versions are not supported
for writing.}

\item{storage_options}{a list of storage options for the zarr array
(e.g. chunks)}

\item{type}{The type of OME pyramid written: 'image' (default) or 'label'.}
}
\description{
Writes an ome image to the zarr path according to ome-zarr specification
}
