% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ome_zarr}
\alias{plot.ome_zarr}
\title{Plot an \code{ome_zarr} object.}
\usage{
\method{plot}{ome_zarr}(x, level = 1, ...)
}
\arguments{
\item{x}{An \code{ome_zarr} object.}

\item{level}{Integer. The scale level to plot. Defaults to \code{1}
(the highest resolution).}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\value{
None
}
\description{
Plot an \code{ome_zarr} object.
}
\examples{
omezarrzip <- system.file("extdata", 
                          "test_ngff_image_v04.ome.zarr.zip", 
                          package = "rome")
dir.create(td <- tempfile())
unzip(omezarrzip, exdir = td)
x <- ome_read(td)
plot(x)
plot(x, 2)
plot(x, all = TRUE)

}
