% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.ome_zarr}
\alias{[.ome_zarr}
\title{Subset an \code{ome-zarr} object}
\usage{
\method{[}{ome_zarr}(x, ...)
}
\arguments{
\item{x}{An \code{ome-zarr} object.}

\item{...}{Indices to subset the \code{ome-zarr} object.}
}
\value{
A subset of an object of \code{ome_zarr} (OME-Zarr) class representing
an image or label pyramid.
}
\description{
Subset operation is applied on all levels of the multiscale \code{ome-zarr}
object. The result is an \code{ome-zarr} object with the same number of levels,
but each level is subsetted according to the provided indices.
}
\details{
The first image is subsetted using the provided indices, and the
resulting dimensions are used to subset the remaining levels, while
conserving the same scaling factor across levels
}
\examples{
omezarrzip <- system.file("extdata", 
                          "test_ngff_image_v04.ome.zarr.zip", 
                          package = "rome")
dir.create(td <- tempfile())
unzip(omezarrzip, exdir = td)
x <- ome_read(td)
y <- x[1:2, 1:5,1:5]
plot(y, level = 2)

}
