% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ome_write.R
\name{ome_write}
\alias{ome_write}
\title{ome_write}
\usage{
ome_write(
  image,
  path = "/",
  axes = NULL,
  scalefactors = c(2, 2, 2, 2),
  version = c("0.4", "0.5"),
  storage_options = NULL
)
}
\arguments{
\item{image}{an n-dimensional (1<n<6) array representing the image data}

\item{path}{the zarr path to write the image to}

\item{axes}{a character vector specifying the axes of the image
(e.g. c("t", "c", "z", "y", "x"))}

\item{scalefactors}{Scale factors to apply to construct a multiscale image.
Importantly, each scale factor is relative to the previous scale factor.
For example, if the scale factors are c(2, 2, 2), the returned multiscale
image will have 4 scales.}

\item{version}{OME-ZARR version (0.4 or 0.5), lower versions are not supported
for writing.}

\item{storage_options}{a list of storage options for the zarr array
(e.g. chunks)}
}
\description{
Writes an ome image to the zarr path according to ome-zarr specification
}
