% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/romeo-package.R
\docType{package}
\name{romeo-package}
\alias{romeo}
\alias{romeo-package}
\title{romeo: Minimal R 'OME-Zarr' Reader}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://huber-group-embl.github.io/romeo/}
  \item \url{https://github.com/Huber-group-EMBL/romeo}
  \item Report bugs at \url{https://github.com/Huber-group-EMBL/romeo/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson@embl.de} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})

Authors:
\itemize{
  \item Hugo Gruson \email{hugo.gruson@embl.de} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})
  \item Artür Manukyan \email{artur-man@hotmail.com} (\href{https://orcid.org/0000-0002-0441-9517}{ORCID})
}

}
\keyword{internal}
